// src/app/generated/caster-api/model/design.ts
/**
 * Caster API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { DesignModule } from './designModule';


export interface Design { 
    id?: string;
    name?: string | null;
    directoryId?: string;
    modules?: Array<DesignModule> | null;
    enabled?: boolean;
}